Fixed the implementation to work with Jackson.
Now field retrieval that expect a number cast to Number instead of Long, and fields expected to be arrays check if it's a List and convert that into an array.
Applications are in general a bit tied to the JSON library used, but with little extra effort they can be portable across JSON libraries, at least Jetty's and Jackson's.