Improved handling of connections back to other comets.
When comet1 is connecting to comet2 with url2A, comet2 connects back to comet1.
It is possible that comet1 has not received yet the handshake response from the first
connection, so it may connect to comet2 with url2B.
This will cause unwanted races among connections, and may result in OortComets
returned by observeComet() to be disconnected.
The fix ensures that comet1 detects if it is the originator of a connection, and if so,
avoid to connect again.