Additional fix for COMETD-317 (JavaScript websocket transport not robust in case of connection failures).
In case of /meta/connect failure, we used to first increase the backoff and then retry; this is now changed
to first retry and then increase the backoff.
The reason is to have quicker reconnections for the websocket transport, in case there is a failure
in a publish that cause the connection to be closed, affecting also an outstanding connect.
In http transports this was not happening because we used two connections.