Fixes COMETD-322 (Massive memory leak during stress testing detected).
When websocket clients abruptly closed the connections, the ServerSession remained in the server for a long period of time (16 minutes).
This was due to incorrectly canceling the interval timeout upon disconnect, so that server thought that a connect was sent,
and waited those 16 minutes before sweeping them.
Now we correctly set the interval timeout, so that the sweeping mechanism can do its job.