Notifying that a comet has left only if the Oort is running.
OortA and OortB are mutually connected; when OortA is stopped, OortCometAB disconnects
and OortB is notified of this. Upon this notification, OortB disconnects OortCometBA and then
notifies the applications that OortA left.
OortA receives the notification that OortCometBA disconnected, but since OortA is stopping, it
should not notify applications that OortB left (because it just disconnected from a stopping Oort).