Avoid dangerous race condition when one method is used simultaneously.
The problem here is that one thread can record the accessible flag
of the method as false, set it to true, then run the method and
reset it to false while a second records the accessible flag as
true and runs the method, which fails, more or less silently.  (An
exception is logged, but under the log4j heirarchy of the calling
service, not cometd).