Reverted fix for COMETD-428 (Enable configuration for "ws.minVersion" from CometdServlet's init-param (web.xml)).
The reason to revert is due to the fact that enabling hixie versions
of the WebSocket protocol exposes to Safari and Jetty bugs.
The browser will fall back to long-polling, so functionality will
still be there, and the server will be safer from bugs.
Also, non-RFC WebSocket versions are not supported by Jetty 9 and
hence CometD 3.