Got rid of the old logging configuration system.
It was based on Jetty loggers, which could be configured at runtime via
methods such as setDebugEnabled(). Unfortunately, not all places in
the code were propagating such settings, leading to inconsistent
logging behavior.

Now the server side relies only on SLF4J, and as such it must be
configured via one of its binding (e.g. Log4J).
Ditto for the Java client.

The JavaScript client still relies on a "logLevel" property that
controls the window.console object.