Fixes COMETD-481 (Advice "multiple-clients" may be generated due to race condition).
LongPollingScheduler.decBrowserId() was called from onComplete(), but it could be
too late: the /meta/connect response has already been sent over the network, and
the client could have issue a new /meta/connect request on another connection
that would be processed concurrently with the finish of the previous one.
This would lead to an incorrect calculation of the browserId count which would
generate the "multiple-clients" advice.

Fixed by moving the decBrowserId() call to onTimeout() so that it is executed
before sending the response over the network.