COMETD-511 - Ack extension may lose messages.
Rewrote the acknowledgement extension (server-side) to behave
correctly in case of non-normal arrival of /meta/connect requests.

Introduced a new listener, ServerSession.QueueListener, that is
notified when a message is added to the session queue.

The new implementation of the acknowledgement extension now
relies on ServerSession.QueueListener and on the existing
ServerSession.DeQueueListener to store messages at the correct time
and to replace the queue at the correct time.