Improved RMI ports handling
RMI Server opens up a serversocket on port 1099 and then listens here
for incoming requests. If a request comes in, ANOTHER port is used to
handle the request and send back the response. In the default
implementation, this is done on port 0. (any available port, chosen
randomly by the underlying OS). The fixed rmiObjectsPort has been
introduced to manage the ports through FW