COMETD-549 - Enable/Disable Jquery Global Ajax Events.
Added "global: false" to the jquery ajax calls.
Global events are not fired for cross-domain or JSONP requests, so
disabling them also for normal calls would make behavior consistent.
Also, if these events are used to show a wait cursor to the user,
this will not work because CometD long polls and therefore the cursor
will always be on.