Do not reject connection messages in the same HTTP request as others.
From the Bayeux specification: " A client MAY send other messages in the same HTTP request with a connection message. A server MUST handle any other message sent in the same request as a connect message after the handling of the connect message is complete."

This implies that connection messages that are sent alongside other messages should not be met with an `IOException`, but instead handled first. This is the behaviour expected by the Faye client. In this commit, this is done by sorting the messages such that the connection messages come first.

We also sort the messages such that any handshake message are seen immediately, so that we can fail quickly.