Fixes #608 (Improve protection against stolen clientId).
HTTP transports now send a HttpOnly cookie that identifies the
Bayeux session, and the implementation maintains a link between
this session cookie and the ServerSession id.

Because WebSocket transports cannot carry metadata, they can only
trust messages on a connection onto which the Bayeux handshake was
performed.

Introduced new parameter "requireHandshakePerConnection" to control
this behavior. To maintain backward compatibility, its value is
false by default, to cover the cases where the connection is broken
and a /meta/connect is send on the new connection (otherwise a new
handshake would be required).

Conflicts:
	cometd-java/cometd-java-server/src/main/java/org/cometd/server/transport/AbstractHttpTransport.java