Issue #604 - Reduce reconnection roundtrips.
The previous implementation was broken in 2 ways.
 - It was executing the Runnable action (introduced to run them
   after the messages that could come with the handshake reply)
   _after_ the state updaters were modified, and this was breaking
   waitFor().
 - It assumed that the messages coming with the handshake reply
   were coming in a single batch with the handshake reply, which
   was true for HTTP transports but not for WebSocket transports.

Introduced field "x-messages" to track the number of messages
sent along with the handshake reply, and properly adding it on
the server and handling it on the client.

Finally, the BayeuxClient implementation has been rewritten using
a queue when a state change was required.