Issue #637 Unsplit packages, websocket common into subpackage
Both the bundle cometd-java-websocket-common-server and
cometd-java-websocket-javax-server export the same package:
org.cometd.websocket.server.

This is called a "split package" in OSGi and its a best practice to
avoid it. It makes using cometd websocket under OSGi difficult.

Move the classes in the common module into a subpackage called "common".

Signed-off-by: Raman Gupta <rocketraman@gmail.com>