Issue #650 - Improve session reconnection detection.
Now the message handling works in this way:
* Upon receiving a message, the expiration timeout is canceled.
* When messages are written to the client, the expiration timeout is
scheduled.

This allows the expiration timeout to be refreshed every time a
message is received, and also allowed to formalize a configuration
parameter called "maxProcessing", which is the max time a message can
take to be processed by the application, by default infinite.