Removed ServerMessageImpl.local property.
This property was introduced as an optimization to
avoid to convert to JSON messages that end their
journey in the server.
However, it was a bit obscure how it worked, since
it was relying on the fact that parsed messages
are not created via BayeuxServer.newMessage().

Replaced the optimization with a check on whether
the message is a reply message, and removed the
ServerMessageImpl.local property.