First pass at making CometD fully async.
Modified the relevant API to accept a Promise, a callback object invoked
when the operation succeeds or fails, in particular message listeners
and extensions.

Reworked the implementation of HTTP transports to handle the additional
asynchronous calls, now stable enough to run the load tests.

Reworked the implementation of the WebSocket transports just enough to
make them compile, but they are not working properly yet.