Issue #737 - Improve performance when using Nashorn for tests (#738)
Instead of engine-per-test, uses a single Nashorn engine, with
bindings-per-test. This allows reuse of already compiled scripts by the
engine. Uses URLReader and ScriptContext.FILENAME to communicate the URL
to the engine, again for reuse. Releases the reference to
AbstractCometD.javaScript in its destroyJavaScript method to make sure
JavaScript's Thread and Bindings are released.

Signed-off-by: Attila Szegedi <szegedia@gmail.com>