Support for a CometD proxy.
This solution uses Message.handled so that applications can write
an extension marks the message as handled but returns false from
the extension handling.

Returning false from the extension tells the implementation that
the extension processing must be interrupted, but the message is
handled and therefore it must not be deleted.