Issue #816 - Normalize server handling of WebSocket close frame.
The handling in CometD 4.0.x is slightly different than CometD 3.1.x.
We need to detach the scheduler from the session because if a message
is delivered on a new WebSocket connection, it may still see the old
scheduler in the session and try to use it to deliver the message.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>