Fixes #890 - Introduce client turnover in benchmark.
Introduced new parameter "clientTurnover" in clients/s.
If greater than zero, it disconnects existing clients
and replaces them with new clients.
This is done at the end of a batch send.

In flight messages for the disconnected clients will be
correctly reported because A) HttpClient always completes the
response (unless the EndPoint is closed) and B) BayeuxClient
does not at its state when there are messages to deliver.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>