Fixes #981 - WebSocketTransport.Factory does not add WebSocketContainer as a bean.
Adding WebSocketContainer as a bean has no effect because WebSocketContainer comes
from the Servlet Container implementation and implements the server version of
LifeCycle, so the webapp version of LifeCycle won't recognize it as a LifeCycle
and won't manage its life.

However, it is useful to add it for dump() purposes.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>