Moved Oort tests to their own module.
This was necessary because maven-surefire-plugin could not figure out
correctly the module-path to run the tests, so they could not be run
in the same module where the source code (src/main/java) is.

In order to preserve the ability to call package/protected methods
from the tests, the tests should remain in the same java package,
which will now generate a split-jar.
However, the split-jar is a test jar that is not deployed to
Maven Central, so it should practically never be a problem,
since it can never end up in a module-path.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>