Fixes #1716 - WebSocket session leak leads to OOME
There was an assumption that a Scheduler was installed only in case of /meta/connect messages.
This is true for HTTP, but not WebSocket.

For the WebSocket transport, the Scheduler may be installed after the /meta/handshake, but before the /meta/connect, to allow server-side messages to be published and reach the client with less latency.
This was happening, for example, if the first message after the /meta/handshake was a /meta/subscribe.

If no /meta/connect message was sent (e.g. due to the connection being closed), the ServerSession was leaked.

The logic in ServerSessionImpl.setScheduler() is changed to properly cancel the expiration without assuming it is a /meta/connect.

Reviewed logic in AbstractWebSocketEndPoint.WebSocketScheduler.schedule() to take into account the fact that the message may not be a /meta/connect.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>
(cherry picked from commit 81946d6c0876b2367af65fb7fccb1bf7074285de)