Fixes #1770 - Suspended sessions are leaked with HTTP/2 reset.
An HTTP/2 reset was causing AsyncListener.onError() to be invoked, and that in turn was calling LongPollScheduler.error(), which was not scheduling the session for expiration.
Even when the CometD timeout fired, trying to write the /meta/connect response resulted in an IllegalStateException that failed to schedule the session for expiration.

Fixed by having AsyncListener.onError() and AbstractReader.onError() call Scheduler.cancel(), and LongPollScheduler.error() call scheduleExpiration(), so that the ServerSession can be swept.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>